fix footer black theme
