+++ src/js/blackTheme/blackTheme.js
@@ -1,22 +1,11 @@
+import blackThemeBody from "./blackThemeBody";
+import blackThemeText from "./blackThemeText";
 localStorage.setItem("toggle", false);
 
 const switcher = e => {
-    console.log(666)
     localStorage.setItem("toggle", e.currentTarget.checked)
-    const textBlack = [...document.querySelectorAll(".film-title")]
-    if (e.currentTarget.checked) {
-        document.body.classList.add('black--body');
-        document.querySelector(".footer").classList.add('black--footer');
-        textBlack.map(elem => elem.classList.add('black--text'));
-        document.querySelector(".switch__checkbox").checked = true
-    }
-    else {
-        document.body.classList.remove('black--body');
-        document.querySelector(".footer").classList.remove('black--footer');
-        textBlack.map(elem => elem.classList.remove('black--text'));
-        document.querySelector(".switch__checkbox").checked = false
-    }
-}
-
+    blackThemeBody();
+    blackThemeText();
+};
 
 export default switcher;
